Move UserInteractiveService to user package
